+++ NEWS.md
@@ -1,4 +1,4 @@
-# vegtable 0.1.9
+# vegtable 0.2.0
 
 ## New Features